requirements.txt
django == 5.0
daphne
channels
djangorestframework == 3.14.0
djangorestframework-simplejwt
django-cors-headers
django-factory_boy